requirements.txt
tensorflow==2.19.0
Pillow==10.3.0
numpy==1.26.4
matplotlib==3.8.4
